📝 chore : 글 작성 페이지 간단한 로직 추가, 예외처리 작업
