+++ src/pages/EditRecruitPostPage/index.js
@@ -1,3 +1,4 @@
+import { postCreatePost, postUploadPostImage } from '@/api/supabase/post_editor';
 import { POSITION_SKILLS } from '@/constants/position';
 
 // 모집 설정
@@ -99,3 +100,69 @@ export const editPositionAndSkills = [
     selectedSkills: [],
   },
 ];
+
+// 데이터 검사 + 전송 로직
+export const sendData = async (userInfo, positionAndSkills) => {
+  let filePath;
+  if (userInfo.post_img_path === '') {
+    // 선택된 이미지 없으면 기본이미지 제공
+    filePath =
+      'https://hfrulsqohffbfxdsozkk.supabase.co/storage/v1/object/public/post_images/programmers.png';
+  } else {
+    filePath = await postUploadPostImage(userInfo.post_img_path);
+  }
+
+  // 최종 전송할 데이터
+  const positions = [];
+  let techStacks = [];
+
+  console.log(userInfo);
+  switch (true) {
+    case userInfo.author === '':
+      return alert('오류 : 유저 로그인 정보가 없습니다.');
+    case userInfo.recruit_type === '':
+      return alert('모집 구분 선택을 완료해주세요');
+    case userInfo.recruit_count === '':
+      return alert('모집 인원 선택을 완료해주세요');
+    case userInfo.recruit_deadline === null:
+      return alert('모집 마감일 선택을 완료해주세요');
+    case userInfo.on_offline === '':
+      return alert('진행 방식 선택을 완료해주세요');
+    case userInfo.start_date === null:
+      return alert('진행 기간 선택을 완료해주세요');
+    case userInfo.end_date === null:
+      return alert('진행 기간 선택을 완료해주세요(종료 날짜가 선택되지 않았습니다)');
+    case userInfo.call_method === '':
+      return alert('연락 방법 선택을 완료해주세요');
+    case userInfo.call_link === '':
+      return alert('연락처 선택을 완료해주세요');
+    case userInfo.title === '':
+      return alert('글 제목 작성을 완료해주세요');
+    case userInfo.body === '':
+      return alert('글 내용 작성을 완료해주세요');
+    default:
+      break;
+  }
+
+  const selected = positionAndSkills.filter((data) => {
+    return data.positionSelected;
+  });
+  if (selected.length === 0) return alert('포지션 선택을 완료해주세요.');
+  for (let i = 0; i < selected.length; i++) {
+    positions.push(selected[i].position);
+    if (selected[i].selectedSkills.length > 0) {
+      techStacks = [...techStacks, ...selected[i].selectedSkills];
+    } else {
+      techStacks = [];
+      return alert('기술스택 선택을 완료해주세요.');
+    }
+  }
+  const setTechStacks = new Set(techStacks);
+  techStacks = [...setTechStacks];
+
+  const resultUserInfo = { ...userInfo, post_img_path: filePath };
+  console.log(resultUserInfo, positions, techStacks);
+
+  await postCreatePost(resultUserInfo, positions, techStacks);
+  alert('전송완료!');
+};
